+++ nhd2-exp/src/gui/epgplus.cpp
@@ -861,7 +861,7 @@ int EpgPlus::exec(CChannelList * _channelList, int selectedChannelIndex, CBouque
 		// add sec timer
 		sec_timer_id = g_RCInput->addTimer(1*1000*1000, false);
 
-		unsigned long long timeoutEnd = CRCInput::calcTimeoutEnd (g_settings.timing[SNeutrinoSettings::TIMING_CHANLIST]);
+		uint64_t timeoutEnd = CRCInput::calcTimeoutEnd (g_settings.timing[SNeutrinoSettings::TIMING_CHANLIST]);
 		bool loop = true;
 
 		while (loop)
