+++ nhd2-exp/src/gui/vfdcontroler.cpp
@@ -121,10 +121,9 @@ int CVfdControler::exec(CMenuTarget* parent, const std::string &)
 
 	setVfd();
 	paint();
-	
 	frameBuffer->blit();
 
-	unsigned long long timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings::TIMING_MENU]);
+	uint64_t timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings::TIMING_MENU]);
 
 	bool loop=true;
 	while (loop)
