+++ nhd2-exp/src/gui/widget/colorchooser.cpp
@@ -185,7 +185,7 @@ int CColorChooser::exec(CMenuTarget *parent, const std::string &)
 
 	int selected = 0;
 
-	unsigned long long timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings::TIMING_MENU]);
+	uint64_t timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings::TIMING_MENU]);
 
 	bool loop = true;
 	while (loop)
